Klokwork: Operands in bitwire opration have different size

+++ fabric_management/daemon/inc/fmd_slave.h
@@ -86,7 +86,7 @@ extern int start_peer_mgmt_slave(uint32_t mast_acc_skt_num, uint32_t mast_did,
 extern void shutdown_slave_mgmt(void);
 
 int add_device_to_dd(ct_t ct, uint32_t did, uint32_t did_sz, hc_t hc,
-                uint32_t is_mast_pt, uint32_t flag, char *name);
+                uint32_t is_mast_pt, uint8_t flag, char *name);
 
 int del_device_from_dd(ct_t ct, uint32_t did);
